Bring more of the Velato runtime, including Composition, Renderer.
This code will need a lot of enhancements and improvements, but it
is still very close to what was in Velato with only minimal changes.

This brings an optional dependency on Vello so that the `Renderer`
can be used to create a Vello `Scene` from the animation and
composition data.